RD-7052 Run configure_manager in RESTservice (#81)
* Don't create a config.yaml where not needed

Those containers only use --db-wait, so they don't need a config file.
What they do need, is prepare_secrets.sh, which sets up the db
credentials.

* api service: use restservice image for the initcontainer

That is the image we in all other places elsewhere for db-wait.
The api-service image won't necessarily always have configure_manager,
but the rest-service image will.
This is similar to what we do in eg. mgmtworker.

* mgmtworker: db-wait && create-admin-token

* restservice: just pass the config file to configure_manager

no need for export or anything.
Also, set the default agent port to 443.

* restservice config initcontainer: also mount ssl

we need the rabbitmq cert to be available!

* restservice: remove internal_rest_port setting

After cloudify-cosmo/cloudify-manager#4073 it is already the default.