implement a server exporting the pod statuses
this is a simple .py HTTP server that on any GET, returns
the status of all pods in the current namespace in the prometheus
metrics format, only giving up{} for each.

This can be a target for prometheus scraping, when we only want
to know if pods are up or not.

It has no dependencies other than the stdlib, so we can run it on
any py3.8+ image.