frontend images: disable runAsNonRoot
Those images, on the 7.0.2 branch, don't set their USER to a numeric
value, but to the username, or don't set it at all. Consequently,
they can't be run while this is in effect.

For the 7.0.2 branch, let's just disable this. On master, the images
are correct, and can be run with runAsNonRoot.