RND-1261 Don't break when agent urls are set to null
If the users sets resources.packages.agents to null, the `curl` loop
would break, because we'd leave only a `&&` rendered, and that's
a bash syntax error.

Instead, let's just not run those containers at all, they're not
needed.

Wrap the `if` expression in parens so that nulling at any level
is valid, e.g. `resources: null`